change readme, add data and type files
