+++ src/types.tsx
@@ -1,0 +1,13 @@
+type PlaceType = {
+  id: number;
+  city: string;
+  src: string;
+  premium: boolean;
+  bookmarks: boolean;
+  price: number;
+  rating: number;
+  description: string;
+  type: string;
+};
+
+export type { PlaceType };
